refactor: 560 - move principles to separate folder
