+++ src/widgets/principles/ui/principles/principles.module.scss
